requirements.txt
alabaster==0.7.9
algoliasearch==1.9.2
algoliasearch-django==1.2.4
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
celery==3.1.24
cffi==1.8.3
cryptography==1.5.2
Django==1.10.2
django-allauth==0.27.0
django-celery==3.1.17
django-filter==0.15.2
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
docutils==0.12
enum34==1.1.6
httplib2==0.9.2
idna==2.1
imagesize==0.7.1
ipaddress==1.0.17
Jinja2==2.8
kombu==3.0.36
Markdown==2.6.7
MarkupSafe==0.23
ndg-httpsclient==0.4.2
oauth2==1.9.0.post1
oauthlib==2.0.0
Pillow==3.4.1
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
PyJWT==1.4.2
pyOpenSSL==16.1.0
python-openid==2.2.5
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.8
sphinx-rtd-theme==0.1.9
